pyproject.toml
[tool.poetry]
name = "pydantic-settings-sources"
version = "0.1.0"
description = "Extra sources for pydantic-settings"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pydantic = "^2.0"
pydantic-settings = "^2.0"
pyyaml = "^6.0"
toml = "^0.10.2"
deepmerge = "^1.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
ruff = "^0.1.0"
tox = "^4.11.1"
black = "^23.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
